refactor(name): recruit schema 파일명 변경 (#76)
